# agent/api.yaml
openapi: "3.0.0"
info:
  version: "0.1.0"
  title: "Squill Agent REST API"
  description: The REST API used by the Squill Client to interact with the Agent.
paths:
...
  /users/{username}/catalog/{catalog_id}/list:
    get:
      summary: List all catalog entries for the specified `username` and `catalog_id`.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Username"
        - name: catalog_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceRef"
        "404":
          description: Catalog ID not found
        "401":
          description: Unauthorized
  /users/{username}/catalog/list:
    get:
      summary: List all root catalog entries for the specified `username`.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Username"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResourceRef"
        "401":
          description: Unauthorized
  /users/{username}/catalog:
    post:
      summary: Create a new resource in the catalog.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: x-resource-type
          in: header
          required: true
          schema:
            enum:
              - connection
              - environment
              - collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/User"
                - $ref: "#/components/schemas/Connection"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceRef"
        "401":
          description: Unauthorized
    put:
      summary: Update an existing resource of the catalog.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: x-resource-type
          in: header
          required: true
          schema:
            enum:
              - connection
              - environment
              - collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/User"
                - $ref: "#/components/schemas/Connection"
      responses:
        "200":
          description: Ok
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /users/{username}/catalog/{catalog_id}/rename:
    post:
      summary: Rename a catalog entry for the specified `username` and `catalog_id`.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Username"
        - name: catalog_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceRef"
        "400":
          description: Invalid Request
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Catalog entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Unprocessable Entity
  /users/{username}/settings:
    post:
      summary: Save the user settings.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Username"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Path not found
        "401":
          description: Unauthorized
  /users/{username}/storage:
    get:
      summary: Get the storage usage of the user.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Username"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStorage"
        "401":
          description: Unauthorized
  /users/{username}/user:
    get:
      parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Username"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
        "401":
          description: Unauthorized
components:
  schemas:
...
    User:
      description: A user that can logon to the system.
      type: object
      required:
        - username
      properties:
        username:
          type: string
...
    Datasource:
      description: The description of a datasource (database for SQL engines).
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: The name of the datasource (e.g. "my_database").
        description:
          type: string
          description: The description of the datasource (e.g. "My database").
        size_in_bytes:
          type: integer
          format: int64
          description: The size of the datasource in bytes.
        hidden:
          type: boolean
          description: A flag indicating if the datasource is hidden in the user interface.
    Connection:
      description: Definition of a connection to datasources.
      type: object
      x-immerable: true
      required:
        - id
        - parent_id
        - owner_user_id
        - mode
        - name
        - save_password
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the connection.
        parent_id:
          type: string
          format: uuid
          description: The unique identifier of the parent resource in the catalog.

        owner_user_id:
          type: string
          format: uuid
          description: The unique identifier of the user that owns the connection.

        driver:
          type: string
          description: The name of the driver used to connect to the datasource.

        name:
          type: string
          description: The name of the connection.

        alias:
          type: string
          format: identifier
          description: The alias of the connection.

        description:
          type: string
          description: The description of the connection.

        mode:
          $ref: "#/components/schemas/ConnectionMode"

        host:
          type: string
          description: The host of the connection.if the connection mode is "host".

        port:
          type: integer
          format: int16
          minimum: 0
          description: The port of the connection.if the connection mode is "host".

        socket:
          type: string
          description: The socket of the connection.if the connection mode is "socket".

        file:
          type: string
          description: The file of the connection.if the connection mode is "file".

        connection_string:
          type: string
          description: The connection string of the connection.if the connection mode is "connection_string".

        uri:
          type: string
          description: The URI of the connection.if the connection mode is "uri".

        username:
          type: string
          description: The username for authentication.

        password:
          type: string
          description: The password for authentication.

        save_password:
          type: boolean
          description: Whether the password should be saved.

        default_datasource:
          type: string
          description: The name of the default datasource to use.

        datasources:
          type: array
          description: The list of datasources available for the connection.
          items:
            $ref: "#/components/schemas/Datasource"

        options:
          type: object
          description: Additional options of the connection.
          additionalProperties:
            type: string
    ConnectionMode:
      description: The mode of a connection.
      type: string
      oneOf:
        - title: host
          description: The connection is a host-based connection.
        - title: socket
          description: The connection is a socket-based connection.
        - title: file
          description: The connection is a file-based connection.
        - title: connection_string
          description: The connection is a connection string-based connection.
        - title: uri
          description: The connection is a URI-based connection.
...
    Error:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          description: The HTTP status code
          type: integer
        code:
          description: A machine readable error code such as `invalid_parameter`.
          type: string
        message:
          description: A human readable error message
          type: string
...
    ResourceType:
      type: string
      oneOf:
        - title: connection
          description: A connection to a datasource.
        - title: environment
          description: An environment that can be used to group connections.
        - title: collection
          description: A collection of resources.
        - title: user
          description: A user.
    ResourceRef:
      description: |
        A reference to a resource.

        A resource reference is a lightweight object that contains the unique identifier of the resource, and some
        additional properties of the resource that are useful to use the resource without having to load it.
      type: object
      properties:
        id:
          description: Unique identifier of the resource.
          type: string
          format: uuid
        parent_id:
          description: Unique identifier of the parent resource.
          type: string
          format: uuid
        owner_user_id:
          description: Unique identifier of the user who own the resource.
          type: string
          format: uuid
        name:
          description: Name of the resource.
          type: string
        type:
          x-rust-name: resource_type
          $ref: "#/components/schemas/ResourceType"
        metadata:
          description: |
            Additional metadata of the resource.

            Metadata are key-value pairs that provide additional information without having to load the resource itself.
            For example, the content type of a `collection`, or the driver of a `connection`.
          type: object
          additionalProperties:
            type: string
      required:
        - id
        - name
        - type
...
    Username:
      description: A username
      type: string
...
    ConnectionStorage:
      description: The storage usage of a connection.
      type: object
      required:
        - connection_id
        - name
        - used_bytes
      properties:
        connection_id:
          description: The unique identifier of the connection.
          type: string
          format: uuid
        name:
          description: The name of the connection.
          type: string
        used_bytes:
          description: The used storage in bytes.
          type: integer
          format: int64
          minimum: 0
    UserStorage:
      description: The storage usage of a user.
      type: object
      required:
        - user_id
        - username
        - connections
      properties:
        user_id:
          description: The unique identifier of the user.
          type: string
          format: uuid
        username:
          description: The username of the user.
          type: string
        connections:
          description: The storage usage of the connections owned by the user.
          type: array
          items:
            $ref: "#/components/schemas/ConnectionStorage"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key

    BearerAuth:
      type: http
      scheme: bearer
security:
  - ApiKeyAuth: []
  - BearerAuth: []